+++ dataiter/__init__.py
@@ -400,6 +400,36 @@ def quantile(x, q, dropna=True):
         return np.nan
     return np.quantile(x, q).item()
 
+def read_csv(path, encoding="utf-8", sep=",", header=True, columns=[], dtypes={}):
+    return DataFrame.read_csv(path,
+                              encoding=encoding,
+                              sep=sep,
+                              header=header,
+                              columns=columns,
+                              dtypes=dtypes)
+
+def read_geojson(path, encoding="utf-8", columns=[], dtypes={}, **kwargs):
+    return GeoJSON.read(path,
+                        encoding=encoding,
+                        columns=columns,
+                        dtypes=dtypes,
+                        **kwargs)
+
+def read_json(path, encoding="utf-8", keys=[], types={}, **kwargs):
+    return ListOfDicts.read_json(path,
+                                 encoding=encoding,
+                                 keys=keys,
+                                 types=types,
+                                 **kwargs)
+
+def read_npz(path, allow_pickle=True):
+    return DataFrame.read_npz(path, allow_pickle=allow_pickle)
+
+read_csv.__doc__ = util.format_alias_doc(read_csv, DataFrame.read_csv)
+read_geojson.__doc__ = util.format_alias_doc(read_geojson, GeoJSON.read)
+read_json.__doc__ = util.format_alias_doc(read_json, ListOfDicts.read_json)
+read_npz.__doc__ = util.format_alias_doc(read_npz, DataFrame.read_npz)
+
 @_ensure_x_type
 def std(x, dropna=True):
     """