+++ dataiter/util.py
@@ -43,6 +43,13 @@ def count_digits(value):
     m = len(parts[1].rstrip("0"))
     return n, m
 
+def format_alias_doc(alias, target):
+    indent = " " * 8
+    note = (
+        ".. note:: :func:`{}` is a convenience alias for :meth:`{}`."
+        .format(alias.__name__, target.__qualname__))
+    return target.__doc__ + "\n\n" + indent + note
+
 def format_floats(seq):
     precision = dataiter.PRINT_FLOAT_PRECISION
     if any(0 < x < 1/10**precision or x > 10**16 - 1 for x in seq):