+++ doc/comparison/blocks/io-binary-dataiter.py
@@ -1,2 +1,2 @@
-data = di.DataFrame.read_npz("data.npz")
+data = di.read_npz("data.npz")
 data.write_npz("data.npz")