+++ doc/comparison/blocks/io-csv-dataiter.py
@@ -1,2 +1,2 @@
-data = di.DataFrame.read_csv("data.csv")
+data = di.read_csv("data.csv")
 data.write_csv("data.csv")